split msgraph into separate package
