+++ llamazure/msgraph/models.py
