+++ llamazure/msgraph/msgraph.py
@@ -8,8 +8,8 @@ from typing import Any, Dict, Optional, Union
 
 import requests
 
-from llamazure.rbac import codec
-from llamazure.rbac.models import Req, Res, ResErr, ResMaybe
+from llamazure.msgraph import codec
+from llamazure.msgraph.models import Req, Res, ResErr, ResMaybe
 
 
 @dataclasses.dataclass
